feat(EIDOMNI-131) : fix tests

+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/api/oid4vci/issuance_v2/CredentialResponseDtoV2.java
@@ -15,6 +15,6 @@ public record CredentialResponseDtoV2(
         String transactionId,
 
         //  REQUIRED if transaction_id is present. Contains a positive number that represents the minimum amount of time in seconds that the Wallet
-        Integer interval
+        Long interval
 ) {
 }
